requirements.txt
PyYAML
fastapi[standard]
pyjwt
passlib[bcrypt]
sqlmodel